Move all terrain section unpacking off main thread.
