+++ src/main/java/grondag/canvas/terrain/region/input/InputRegion.java
@@ -20,9 +20,49 @@
 
 package grondag.canvas.terrain.region.input;
 
-import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EXTERIOR_STATE_COUNT;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_INDEX_000;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_INDEX_002;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_INDEX_020;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_INDEX_022;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_INDEX_200;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_INDEX_202;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_INDEX_220;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_INDEX_222;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_I_MASK;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_J_MASK;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_J_SHIFT;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_K_SHIFT;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.CORNER_STATE_COUNT;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_INDEX_Y0X0;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_INDEX_Y0X2;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_INDEX_Y2X0;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_INDEX_Y2X2;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_INDEX_Z0X0;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_INDEX_Z0X2;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_INDEX_Z0Y0;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_INDEX_Z0Y2;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_INDEX_Z2X0;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_INDEX_Z2X2;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_INDEX_Z2Y0;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_INDEX_Z2Y2;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_I_MASK;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_J_MASK;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_J_SHIFT;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_K_SHIFT;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.EDGE_STATE_COUNT;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.FACE_I_MASK;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.FACE_J_MASK;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.FACE_J_SHIFT;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.FACE_K_SHIFT;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.FACE_STATE_COUNT;
 import static grondag.canvas.terrain.util.RenderRegionStateIndexer.INTERIOR_STATE_COUNT;
 import static grondag.canvas.terrain.util.RenderRegionStateIndexer.REGION_PADDING;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.SIDE_INDEX_X0;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.SIDE_INDEX_X2;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.SIDE_INDEX_Y0;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.SIDE_INDEX_Y2;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.SIDE_INDEX_Z0;
+import static grondag.canvas.terrain.util.RenderRegionStateIndexer.SIDE_INDEX_Z2;
 import static grondag.canvas.terrain.util.RenderRegionStateIndexer.TOTAL_STATE_COUNT;
 import static grondag.canvas.terrain.util.RenderRegionStateIndexer.interiorIndex;
 import static grondag.canvas.terrain.util.RenderRegionStateIndexer.regionIndexToPackedSectionPos;
@@ -40,8 +80,10 @@ import net.minecraft.world.level.BlockAndTintGetter;
 import net.minecraft.world.level.ColorResolver;
 import net.minecraft.world.level.LightLayer;
 import net.minecraft.world.level.biome.Biome;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.chunk.LevelChunkSection;
 import net.minecraft.world.level.lighting.LevelLightEngine;
 import net.minecraft.world.level.material.FluidState;
 
@@ -124,11 +166,125 @@ public class InputRegion extends AbstractInputRegion implements BlockAndTintGett
 			}
 		}
 
-		System.arraycopy(packedRegion.states, 0, states, INTERIOR_STATE_COUNT, EXTERIOR_STATE_COUNT);
+		captureCorners();
+		captureEdges();
+		captureFaces();
 
 		copyBeData(packedRegion);
 
 		occlusion.prepare();
+	}
+
+	private interface BlockStateFunction {
+		BlockState apply (int i, int j, int k);
+	}
+
+	private static final BlockState AIR = Blocks.AIR.defaultBlockState();
+	private static BlockStateFunction AIR_FUNCTION = (i, j, k) -> AIR;
+
+	//NB: the addressing math here must match what is in RenderRegionAddressHelper
+	private void captureFace(int baseIndex, BlockStateFunction func) {
+		for (int n = 0; n < FACE_STATE_COUNT; ++n) {
+			states[baseIndex + n] = func.apply(n & FACE_I_MASK, (n >> FACE_J_SHIFT) & FACE_J_MASK, n >> FACE_K_SHIFT);
+		}
+	}
+
+	private void captureFaces() {
+		final LevelChunkSection lowX = getSection(0, 1, 1);
+		captureFace(SIDE_INDEX_X0, lowX == null ? AIR_FUNCTION : (i, j, k) -> lowX.getBlockState(14 + k, i, j));
+
+		final LevelChunkSection highX = getSection(2, 1, 1);
+		captureFace(SIDE_INDEX_X2, highX == null ? AIR_FUNCTION : (i, j, k) -> highX.getBlockState(k, i, j));
+
+		final LevelChunkSection lowZ = getSection(1, 1, 0);
+		captureFace(SIDE_INDEX_Z0, lowZ == null ? AIR_FUNCTION : (i, j, k) -> lowZ.getBlockState(i, j, 14 + k));
+
+		final LevelChunkSection highZ = getSection(1, 1, 2);
+		captureFace(SIDE_INDEX_Z2, highZ == null ? AIR_FUNCTION : (i, j, k) -> highZ.getBlockState(i, j, k));
+
+		final LevelChunkSection lowY = getSection(1, 0, 1);
+		captureFace(SIDE_INDEX_Y0, lowY == null ? AIR_FUNCTION : (i, j, k) -> lowY.getBlockState(i, 14 + k, j));
+
+		final LevelChunkSection highY = getSection(1, 2, 1);
+		captureFace(SIDE_INDEX_Y2, highY == null ? AIR_FUNCTION : (i, j, k) -> highY.getBlockState(i, k, j));
+	}
+
+	//NB: the addressing math here must match what is in RenderRegionAddressHelper
+	private void captureEdge(int baseIndex, BlockStateFunction func) {
+		for (int n = 0; n < EDGE_STATE_COUNT; ++n) {
+			states[baseIndex + n] = func.apply(n & EDGE_I_MASK, (n >> EDGE_J_SHIFT) & EDGE_J_MASK, n >> EDGE_K_SHIFT);
+		}
+	}
+
+	private void captureEdges() {
+		final LevelChunkSection aaZ = getSection(0, 0, 1);
+		captureEdge(EDGE_INDEX_Y0X0, aaZ == null ? AIR_FUNCTION : (i, j, k) -> aaZ.getBlockState(14 + i, 14 + j, k));
+
+		final LevelChunkSection abZ = getSection(0, 2, 1);
+		captureEdge(EDGE_INDEX_Y2X0, abZ == null ? AIR_FUNCTION : (i, j, k) -> abZ.getBlockState(14 + i, j, k));
+
+		final LevelChunkSection baZ = getSection(2, 0, 1);
+		captureEdge(EDGE_INDEX_Y0X2, baZ == null ? AIR_FUNCTION : (i, j, k) -> baZ.getBlockState(i, 14 + j, k));
+
+		final LevelChunkSection bbZ = getSection(2, 2, 1);
+		captureEdge(EDGE_INDEX_Y2X2, bbZ == null ? AIR_FUNCTION : (i, j, k) -> bbZ.getBlockState(i, j, k));
+
+		final LevelChunkSection aYa = getSection(0, 1, 0);
+		captureEdge(EDGE_INDEX_Z0X0, aYa == null ? AIR_FUNCTION : (i, j, k) -> aYa.getBlockState(14 + i, k, 14 + j));
+
+		final LevelChunkSection aYb = getSection(0, 1, 2);
+		captureEdge(EDGE_INDEX_Z2X0, aYb == null ? AIR_FUNCTION : (i, j, k) -> aYb.getBlockState(14 + i, k, j));
+
+		final LevelChunkSection bYa = getSection(2, 1, 0);
+		captureEdge(EDGE_INDEX_Z0X2, bYa == null ? AIR_FUNCTION : (i, j, k) -> bYa.getBlockState(i, k, 14 + j));
+
+		final LevelChunkSection bYb = getSection(2, 1, 2);
+		captureEdge(EDGE_INDEX_Z2X2, bYb == null ? AIR_FUNCTION : (i, j, k) -> bYb.getBlockState(i, k, j));
+
+		final LevelChunkSection Xaa = getSection(1, 0, 0);
+		captureEdge(EDGE_INDEX_Z0Y0, Xaa == null ? AIR_FUNCTION : (i, j, k) -> Xaa.getBlockState(k, 14 + i, 14 + j));
+
+		final LevelChunkSection Xab = getSection(1, 0, 2);
+		captureEdge(EDGE_INDEX_Z2Y0, Xab == null ? AIR_FUNCTION : (i, j, k) -> Xab.getBlockState(k, 14 + i, j));
+
+		final LevelChunkSection Xba = getSection(1, 2, 0);
+		captureEdge(EDGE_INDEX_Z0Y2, Xba == null ? AIR_FUNCTION : (i, j, k) -> Xba.getBlockState(k, i, 14 + j));
+
+		final LevelChunkSection Xbb = getSection(1, 2, 2);
+		captureEdge(EDGE_INDEX_Z2Y2, Xbb == null ? AIR_FUNCTION : (i, j, k) -> Xbb.getBlockState(k, i, j));
+	}
+
+	//NB: the addressing math here must match what is in RenderRegionAddressHelper
+	private void captureCorner(int baseIndex, BlockStateFunction func) {
+		for (int n = 0; n < CORNER_STATE_COUNT; ++n) {
+			states[baseIndex + n] = func.apply(n & CORNER_I_MASK, (n >> CORNER_J_SHIFT) & CORNER_J_MASK, n >> CORNER_K_SHIFT);
+		}
+	}
+
+	private void captureCorners() {
+		final LevelChunkSection xyz = getSection(0, 0, 0);
+		captureCorner(CORNER_INDEX_000, xyz == null ? AIR_FUNCTION : (i, j, k) -> xyz.getBlockState(14 + i, 14 + j, 14 + k));
+
+		final LevelChunkSection xyZ = getSection(0, 0, 2);
+		captureCorner(CORNER_INDEX_200, xyZ == null ? AIR_FUNCTION : (i, j, k) -> xyZ.getBlockState(14 + i, 14 + j, k));
+
+		final LevelChunkSection xYz = getSection(0, 2, 0);
+		captureCorner(CORNER_INDEX_020, xYz == null ? AIR_FUNCTION : (i, j, k) -> xYz.getBlockState(14 + i, j, 14 + k));
+
+		final LevelChunkSection xYZ = getSection(0, 2, 2);
+		captureCorner(CORNER_INDEX_220, xYZ == null ? AIR_FUNCTION : (i, j, k) -> xYZ.getBlockState(14 + i, j, k));
+
+		final LevelChunkSection Xyz = getSection(2, 0, 0);
+		captureCorner(CORNER_INDEX_002, Xyz == null ? AIR_FUNCTION : (i, j, k) -> Xyz.getBlockState(i, 14 + j, 14 + k));
+
+		final LevelChunkSection XyZ = getSection(2, 0, 2);
+		captureCorner(CORNER_INDEX_202, XyZ == null ? AIR_FUNCTION : (i, j, k) -> XyZ.getBlockState(i, 14 + j, k));
+
+		final LevelChunkSection XYz = getSection(2, 2, 0);
+		captureCorner(CORNER_INDEX_022, XYz == null ? AIR_FUNCTION : (i, j, k) -> XYz.getBlockState(i, j, 14 + k));
+
+		final LevelChunkSection XYZ = getSection(2, 2, 2);
+		captureCorner(CORNER_INDEX_222, XYZ == null ? AIR_FUNCTION : (i, j, k) -> XYZ.getBlockState(i, j, k));
 	}
 
 	private void copyBeData(PackedInputRegion protoRegion) {
